+++ storage/storagemetrics/decorator.go
@@ -1,23 +1,24 @@
 // Copyright (c) 2024 The Jaeger Authors.
 // SPDX-License-Identifier: Apache-2.0
 
-package storage
+package storagemetrics
 
 import (
 	"go.uber.org/zap"
 
 	"github.com/jaegertracing/jaeger/pkg/metrics"
+	"github.com/jaegertracing/jaeger/storage"
 	"github.com/jaegertracing/jaeger/storage/dependencystore"
 	"github.com/jaegertracing/jaeger/storage/spanstore"
 	"github.com/jaegertracing/jaeger/storage/spanstore/spanstoremetrics"
 )
 
 type DecoratorFactory struct {
-	delegate       Factory
+	delegate       storage.Factory
 	metricsFactory metrics.Factory
 }
 
-func NewDecoratorFactory(f Factory, mf metrics.Factory) *DecoratorFactory {
+func NewDecoratorFactory(f storage.Factory, mf metrics.Factory) *DecoratorFactory {
 	return &DecoratorFactory{
 		delegate:       f,
 		metricsFactory: mf,
