+++ storage/storagemetrics/decorator_test.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2024 The Jaeger Authors.
 // SPDX-License-Identifier: Apache-2.0
 
-package storage
+package storagemetrics
 
 import (
 	"context"
